\name{ckheartRate}
\alias{ckheartRate}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Data collected from 9/12/2017 to 9/28/2017
}
\description{
Data collected on resting heart rate as well as miles walked, time spent on homework (in each class and overall), and estimated amount of sleep
}
\usage{
ckheartRate
}
%- maybe also 'usage' for other objects documented here.


\value{
one dataset contain 21 observations over 9 variables.
}

\author{
Corey Kimzey
}


\seealso{
\code{\link{updating_heartRate}}
}
\examples{
data(ckheartRate)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ heart }% use one of  RShowDoc("KEYWORDS")

