\name{updating_heartRate}
\alias{updating_heartRate}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Resting Heart Rate updating
}
\description{
Updates heart rate data from my google spread sheet
}
\usage{
updating_heartRate()
}
\value{
Will return a dataset
}

\author{
Corey Kimzey
}
\note{
Will update with the most current data
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
 \code{\link{ckheartRate}}
}
\examples{
#will not run if don't have access to file

#testing <- updating_heartRate()

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ updating}% use one of  RShowDoc("KEYWORDS")

